fix(export-secrets): fix a bug
