+++ export-secrets/package.json
@@ -5,7 +5,8 @@
   },
   "dependencies": {
     "@actions/core": "1.11.1",
-    "js-yaml": "4.1.0"
+    "js-yaml": "4.1.0",
+    "lib": "file:../lib/dist"
   },
   "devDependencies": {
     "@types/js-yaml": "4.0.9",
