+++ get-target-config/src/run.test.ts
@@ -2,7 +2,10 @@ import { run, Result } from "./run";
 
 test("default", async () => {
   const result: Result = {
-    envs: new Map<string, any>([["TFACTION_WORKING_DIR", "tests/aws/foo"]]),
+    envs: new Map<string, any>([
+      ["TFACTION_WORKING_DIR", "tests/aws/foo"],
+      ["TFACTION_TARGET", "tests/aws/foo"],
+    ]),
     outputs: new Map<string, any>([
       ["working_directory", "tests/aws/foo"],
       [
@@ -43,7 +46,10 @@ test("default", async () => {
 
 test("config", async () => {
   const result: Result = {
-    envs: new Map<string, any>([["TFACTION_WORKING_DIR", "tests/aws/foo"]]),
+    envs: new Map<string, any>([
+      ["TFACTION_WORKING_DIR", "tests/aws/foo"],
+      ["TFACTION_TARGET", "tests/aws/foo"],
+    ]),
     outputs: new Map<string, any>([
       ["working_directory", "tests/aws/foo"],
       [
@@ -95,7 +101,10 @@ test("config", async () => {
 
 test("scaffold_working_dir", async () => {
   const result: Result = {
-    envs: new Map<string, any>([["TFACTION_WORKING_DIR", "tests/aws/foo"]]),
+    envs: new Map<string, any>([
+      ["TFACTION_WORKING_DIR", "tests/aws/foo"],
+      ["TFACTION_TARGET", "tests/aws/foo"],
+    ]),
     outputs: new Map<string, any>([
       ["working_directory", "tests/aws/foo"],
       [
