+++ lib/package.json
@@ -8,12 +8,13 @@
   },
   "dependencies": {
     "@actions/core": "1.11.1",
+    "@actions/exec": "^1.1.1",
     "js-yaml": "4.1.0",
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "typescript": "5.6.3",
     "@types/js-yaml": "4.0.9",
-    "@types/node": "20.16.11"
+    "@types/node": "20.16.11",
+    "typescript": "5.6.3"
   }
 }
